Error in formula corrected
The formula in the Powerpoint and the final Word doc was wrong - the
sign of x-beta was incorrect.  This mistake carried into the
calculations immediately after it in the R Markdown doc, so the
probabilities being calculated in those two paragraphs were incorrect.
The rest of the paper's probabilties are fine, as they used my P.X
function from perfman, which did not have this error